+++ packages/apptive_grid_core/lib/src/model/data_entity.dart
@@ -91,6 +91,10 @@ abstract class DataEntity<T, S> with FilterableMixin {
         );
       case DataType.uri:
         return UriDataEntity.fromJson(json);
+      case DataType.email:
+        return EmailDataEntity(json);
+      case DataType.phoneNumber:
+        return PhoneNumberDataEntity(json);
     }
   }
 }
@@ -497,4 +501,22 @@ class UriDataEntity extends DataEntity<Uri, String> {
   /// Returns [value] as a json object map
   @override
   String? get schemaValue => value?.toString();
+}
+
+/// [DataEntity] representing an email
+class EmailDataEntity extends DataEntity<String, String> {
+  /// Creates a new StringDataEntity Object with value [value]
+  EmailDataEntity([super.value]);
+
+  @override
+  String? get schemaValue => value;
+}
+
+/// [DataEntity] representing an email
+class PhoneNumberDataEntity extends DataEntity<String, String> {
+  /// Creates a new StringDataEntity Object with value [value]
+  PhoneNumberDataEntity([super.value]);
+
+  @override
+  String? get schemaValue => value;
 }