+++ packages/apptive_grid_core/test/data_entity_test.dart
@@ -305,4 +305,34 @@ void main() {
       expect(entity.schemaValue, isNull);
     });
   });
+
+  group('Email', () {
+    test('Value is set', () {
+      final entity = EmailDataEntity('Value');
+
+      expect(entity.value, equals('Value'));
+      expect(entity.schemaValue, equals('Value'));
+    });
+
+    test('Default is null', () {
+      final entity = EmailDataEntity();
+
+      expect(entity.value, equals(null));
+    });
+  });
+
+  group('Phone Number', () {
+    test('Value is set', () {
+      final entity = PhoneNumberDataEntity('123456789');
+
+      expect(entity.value, equals('123456789'));
+      expect(entity.schemaValue, equals('123456789'));
+    });
+
+    test('Default is null', () {
+      final entity = PhoneNumberDataEntity();
+
+      expect(entity.value, equals(null));
+    });
+  });
 }